Add driver for  Sensirion SHT3x-D Humidity and Temperature Sensor.
This only add support for base functionality, but it's perfectly
functional as is.

I left in the debugging support used for developing the driver as I'm
fairly certain it will be useful for futre debugging and enhancements.
I think having a standard way of leaving in debugging support for
drivers would be useful, especially as the log subsystem elegantly
supports tuning how much to output and it can be compiled out by
default, but it may be there are better ways of doing this.