Change mqtt_client keep alive from 10s to 5m.
In the current mqtt_client, when the keep alive expires there is no
way to reconnect, so there is no advantage in a short keep alive
period as mqtt_client just silently stops working at that point.

This is mostly a bandaid until larger rework of connection options is
done for #109, #110 and #113. Since that rework is required, this
change is intentionally kept simple and does not allow user override
of the keep alive interval.

While here use BTASSERTI in a number of cases to make test output more
usable.

This was tested using mqtt_client, and mqtt_client_network against a
real MQTT broker (Mosquitto).