Use i2c module instead of i2c_soft in SHT3XD driver.
Unfortunately, this forces boards with hardware I2C to use a hardware
I2C interface with this driver. However, I think it's better than
forcing users to use a software I2C implementation when hardware I2C
is available. Could add a level of indirection with function pointers
to support both, but that's to be implemented later (if needed).

NOTE: This commit has not been tested since I don't have the required
      hardware.