Check TS and RS bits of Status register in can_start(...)
If CAN-RX pin is not connected to transceiver or in case
any other errors in CAN-bus it is better to know about as early
as possible.

can_start(...) returns -ENETDOWN in this case.

At start these bits are set beacuse CAN-module "is waiting
for 11 consecutive recessive bits". Therefore we are polling
these bits in a cycle but not longer than is needed for a frame
of maximal length at speed 10 KBPS