CAN-bus errors and writing thread rapid resuming (#130)
* ESP32 CAN reset on error and resume writer thread.

* Weaker conditions of resetting the hardware and writing thread in case of error interrupts.
Thread waiting for write(...) finish is awaken and hardware is reset only
if bits BS (7, Bus Status) or ES (6, Error Status) are set. The specific
source of interrupt is not analyzed as different situations may generate
different interrupts.

* Fixing the situation when hardware is started with CAN-RX disconnected from trasceiver.
In this situation bits TS (5, Transmit Status) and RS (4, Receive Status)
of Status register are set. According to docs for SJA1000 that means
"controller is waiting to become idle again" (see note 3 for Status register
description). In this situation writing to hardware is uselessly.

* Accelerating the resuming of the thread waiting for write(...) finish.
Without this code the thread will be resumed only in 10 ms, at next
FreeRTOS's time slice.
This is copy-paste from uart_port.i.
TODO: make the same changes for the rest of the periphery.

* Check TS and RS bits of Status register in can_start(...)

If CAN-RX pin is not connected to transceiver or in case
any other errors in CAN-bus it is better to know about as early
as possible.

can_start(...) returns -ENETDOWN in this case.

At start these bits are set beacuse CAN-module "is waiting
for 11 consecutive recessive bits". Therefore we are polling
these bits in a cycle but not longer than is needed for a frame
of maximal length at speed 10 KBPS

* Code cleaned and semaphore is given unconditionally

- Many source formatting issues are solved
- New static function 'is_waiting_for_bus_idle(...)' is
writed for readability
- Semaphore is given unconditionally at end on interrupt handler
- (-ENETDOWN) instead of (-EIO) is returned from write_cb in case
module is waiting for bus idle

* Refactoring: is_waiting_for_bus_idle(), wait_for_bus_idle()

- Changed argument type of 'is_waiting_for_bus_idle(...)'
Now it accepts the value of STATUS register

- New function 'wait_for_bus_idle(...)' extracted from 'can_port_start(...)'

* BUGFIX: the right time control

* Change arguments type of internal funcions and clean code.

- functions wait_for_bus_idle(...) and reset_hw(...) are internal and
need access to hardware registers only
- function wait_for_bus_idle(...) moved up to the function is_waiting_for_bus_idle(...)
- removed superfluous spaces

* Remove variable still_waiting_for_bus_idle from wait_for_bus_idle(...).

* Code style changes