mqtt_client: Add support for connection options.
This allows setting client id, username, password, and Last Will and
Testament in MQTT connection options.

Setting this is done through a parameter to
mqtt_client_connect(). While this changes the API, it seems like the
best aproach as it allows the client not to have to keep any state for
the new options and therefor do not increase memory consumption other
than during connect.

The change is made so it's very easy for existing code to be updated,
as it simply need to change connect call from:

  mqtt_client_connect(client);

to:

  mqtt_client_connect(client, NULL);

to preserve existing behaviour.

While here update the documentation.

This has been tested against real MQTT brokers io.adafruit.com and
test.mosquitto.org.